Refactored DerivedViewInline - moved special fields to separate files
