+++ frontend/src/common/views/fields/inline/DerivedViewInline.jsx
@@ -1,33 +1,11 @@
 import React from 'react';
 
-import {
-  Col, Container, Form, Row,
-} from 'react-bootstrap';
+import { Container, Row } from 'react-bootstrap';
+import InputForm from './derived/InputForm';
+import PersonalInfoField from './derived/PersonalInfoField';
 
-import dataValidator from '../../../helpers/DataValidator';
 import derivedHelper from '../../../helpers/DerivedHelper';
 
-
-const InputForm = ({
-  title, text, setText, isInvalid,
-  isBlocked,
-}) => {
-  const dirty = text.length > 0;
-  return (
-    <Form onSubmit={(e) => e.preventDefault()}>
-      <Form.Control
-        as="input"
-        size="lg"
-        rows={1}
-        value={text}
-        placeholder={title}
-        onChange={(event) => setText(event.target.value)}
-        isInvalid={dirty && isInvalid}
-        disabled={isBlocked}
-      />
-    </Form>
-  );
-};
 
 const OneField = ({
   derivedType, index,
@@ -36,61 +14,14 @@ const OneField = ({
   isBlocked,
 }) => {
   if (derivedType === 'BIRTHDAY_PESEL' && index === 0) {
-    const textAndValidator = input[index] ? JSON.parse(input[index]) : { type: 'PESEL', value: input[index] };
-    const text = textAndValidator.value;
-    const fieldValidator = textAndValidator.type;
-
-    const validators = ['PESEL', 'Nr dow. os.', 'Nr paszportu', 'Inne'];
-    const options = validators
-      .map((o, i) => (<option key={o} value={validators[i]}>{o}</option>));
-
-    const setNewInput = (value) => {
-      const newInput = input.slice();
-      newInput[index] = value;
-      const newValues = fieldValidator === 'PESEL' ? derivedHelper.calculateDerived(derivedType, index, newInput) : newInput;
-      newValues[index] = JSON.stringify({ type: fieldValidator, value });
-      setInput(newValues);
-    };
-
-    const setValidator = (validator) => {
-      const newInput = input.slice();
-      const newValues = validator === 'PESEL' ? derivedHelper.calculateDerived(derivedType, index, newInput) : newInput;
-      newValues[index] = JSON.stringify({ type: validator, value: text });
-      setInput(newValues);
-    };
-
-    const isValid = dataValidator.validateDerivedField(input[index], index, derivedType);
     return (
-      <>
-        <Col>
-          <div className="w-100">
-            <InputForm
-              title={fieldValidator}
-              text={text}
-              setText={setNewInput}
-              isInvalid={!isValid}
-              isBlocked={isBlocked}
-            />
-          </div>
-        </Col>
-        <Col sm="auto">
-          <div className="w-100">
-            <Form>
-              <Form.Group>
-                <Form.Control
-                  as="select"
-                  size="lg"
-                  onChange={(event) => setValidator(event.target.value)}
-                  value={fieldValidator}
-                  disabled={isBlocked}
-                >
-                  {options}
-                </Form.Control>
-              </Form.Group>
-            </Form>
-          </div>
-        </Col>
-      </>
+      <PersonalInfoField
+        input={input}
+        setInput={setInput}
+        index={index}
+        derivedType={derivedType}
+        isBlocked={isBlocked}
+      />
     );
   }
